Fixed progressive decoding without subbanddiff
(cherry picked from commit e3445eefab77c6666d760fa042f141c0a2d2965a)

+++ server/shadow/shadow_client.c
@@ -927,11 +927,11 @@ static BOOL shadow_client_send_surface_bits(rdpShadowClient* client, BYTE* pSrcD
                                             int nXSrc, int nYSrc, int nWidth, int nHeight)
 {
 	BOOL ret = TRUE;
-	int i;
+	size_t i;
 	BOOL first;
 	BOOL last;
 	wStream* s;
-	int numMessages;
+	size_t numMessages;
 	UINT32 frameId = 0;
 	rdpUpdate* update;
 	rdpContext* context = (rdpContext*)client;
